pyproject.toml
[project]
name = "conta-azul"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.2",
    "db-dtypes>=1.4.4",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-secret-manager>=2.25.0",
    "pandas>=2.3.3",
    "pandas-gbq>=0.31.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
